chore: ping is now a function
