+++ src/ping/README.md
@@ -8,16 +8,11 @@ libp2p-ping JavaScript Implementation
 ## Usage
 
 ```javascript
-var Ping = require('libp2p-ping')
+var Ping = require('libp2p/src/ping')
 
-Ping.mount(swarm) // Enable this peer to echo Ping requests
+Ping.mount(libp2p) // Enable this peer to echo Ping requests
 
-var p = new Ping(swarm, peerDst) // Ping peerDst, peerDst must be a peer-info object
+const latency = await Ping(libp2p, peerDst)
 
-p.on('ping', function (time) {
-  console.log(time + 'ms')
-  p.stop() // stop sending pings
-})
-
-p.start()
+Ping.unmount(libp2p)
 ```
